Avances en crear anuncio y arreglos en facade
Al cambiar el nombre user por nick en la BBDD tuve que cambiarlo tambien en las consultas de facade.php
Además he añadido nuevas funciones en facade para tratar anuncios, categorias e imagenes

+++ bbdd/createTable.sql
@@ -41,7 +41,7 @@ CREATE TABLE final_anuncio(
 	idUser INT(10) UNSIGNED NOT NULL,
 	idCategoria INT(5) UNSIGNED DEFAULT 0,
     fecha DATETIME DEFAULT CURRENT_TIMESTAMP,
-    precio INT(10) DEFAULT 0,
+    precio DECIMAL(7,2) DEFAULT 0,
     titulo VARCHAR(50) NOT NULL,
     descripcion VARCHAR(2000),
     localizacion VARCHAR(50) NOT NULL,
